requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-appconf==1.0.2
django-bootstrap4==0.0.7
django-ckeditor==5.6.1
django-heroku==0.3.1
django-import-export==1.1.0
django-js-asset==1.1.0
easy-thumbnails==2.5
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
odfpy==1.4.0
openpyxl==2.5.12
Pillow==5.3.0
pinax-messages==2.0.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.10.11
xlrd==1.2.0
xlwt==1.3.0
